WIP: try binary string version of mixed integer region assignment
Previously, I have always used a list of binary variables to encode
region membership. For example, if we have 8 regions, then we have 8
binary variables per step, with b_i == 1 => membership in region i

This adds a new formulation with only log(num_regions) binary variables
per step. The entire string of variables now encodes the particular
region in which a step is located. For example, with 8 regions, we would
have binary region assignments:

000 => in region 0
001 => in region 1
010 => in region 2
...
111 => in region 7

This works, but so far it's been slower than my previous formulations